Add a seed data class and script command redelegated random seed data into its own class and build the script command
