+++ package.json
@@ -9,7 +9,8 @@
   },
   "scripts": {
     "react-dev": "webpack -d --watch",
-    "server-dev": "nodemon server/server.js"
+    "server-dev": "nodemon server/server.js",
+    "seed-data": "node server/database/SeedData.js"
   },
   "dependencies": {
     "babel-core": "^6.26.0",
